Add parcel build for development mode
