+++ src/.postcssrc
@@ -1,0 +1,6 @@
+{
+  "plugins": {
+    "postcss-import": {},
+    "tailwindcss": {}
+  }
+}
