+++ src/components.js
@@ -1,0 +1,2 @@
+import './styles/index.css';
+
