+++ src/package.json
@@ -2,10 +2,20 @@
   "name": "components",
   "version": "1.0.0",
   "description": "Liquid component library module for platformOS",
-  "main": "index.js",
+  "targets": {
+    "default": {
+      "source": "./components.js",
+      "distDir": "../public/assets"
+    },
+    "styleguide": {
+      "source": "./styleguide.js",
+      "distDir": "../public/assets"
+    }
+  },
+  "browserslist": "> 0.5%, last 2 versions, not dead",
   "scripts": {
-    "dev": "tailwindcss -i ./styles/index.css -o ../public/assets/components.css --watch",
-    "build": "tailwindcss -i ./styles/index.css -o ../public/assets/components.css --minify"
+    "start": "parcel watch --host localhost",
+    "build": "parcel build --no-source-maps"
   },
   "repository": {
     "type": "git",
@@ -18,6 +28,9 @@
   },
   "homepage": "https://github.com/Platform-OS/pos-module-components#readme",
   "devDependencies": {
+    "parcel": "^2.7.0",
+    "postcss": "^8.4.16",
+    "postcss-import": "^14.1.0",
     "tailwindcss": "^3.1.8"
   }
 }
