+++ src/styles/elements.css
@@ -1,0 +1,5 @@
+@layer components {
+  hr:not(.pos-clean) {
+    @apply my-4 border-t-divider;
+  }
+}
